TDQ-6166 avoid tdqRepService null.
git-svn-id: http://talendforge.org/svn/tos/trunk@92341 f6f1c999-d317-4740-80b0-e6d1abc6f99e

+++ org.talend.core.repository/src/main/java/org/talend/repository/ui/actions/DeleteAction.java
@@ -1275,7 +1275,9 @@ public class DeleteAction extends AContextualAction {
         if (GlobalServiceRegister.getDefault().isServiceRegistered(ITDQRepositoryService.class)) {
             ITDQRepositoryService tdqRepService = (ITDQRepositoryService) GlobalServiceRegister.getDefault().getService(
                     ITDQRepositoryService.class);
-            tdqRepService.removeAliasInSQLExplorer(node);
+            if (tdqRepService != null) {
+                tdqRepService.removeAliasInSQLExplorer(node);
+            }
         }
     }
